// src/config.d.ts
export declare const config: {
    discordToken: string;
    clientId: string;
    newsChannelId: string;
    countdownChannelId: string;
    rssFeeds: {
        name: string;
        url: string;
    }[];
    countdownEvents: {
        name: string;
        date: Date;
        description: string;
    }[];
};
//# sourceMappingURL=config.d.ts.map